requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.5.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pydantic==2.2.1
pydantic_core==2.6.1
pytest==7.4.0
requests==2.31.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
